Renamed controller file
I noticed that the class definition for the Alienware 13 R3 controller had been incorrectly placed in the file named controller_m13xr3.py. I found this confusing when debugging the project on my 13 R3 and have renamed the file to correctly reflect the class definition it contains.
