+++ alienfx/core/prober.py
@@ -45,7 +45,6 @@ import alienfx.core.controller_m11xr1
 import alienfx.core.controller_m11xr2
 import alienfx.core.controller_m11xr3
 import alienfx.core.controller_m13xr2
-import alienfx.core.controller_m13xr3
 import alienfx.core.controller_m14xr1
 import alienfx.core.controller_m14xr2
 import alienfx.core.controller_m14xr3
@@ -53,6 +52,7 @@ import alienfx.core.controller_m15x
 import alienfx.core.controller_m17x
 import alienfx.core.controller_m17xr3
 import alienfx.core.controller_m17xr4
+import alienfx.core.controller_13r3
 import alienfx.core.controller_17r1
 import alienfx.core.controller_17r3
 import alienfx.core.controller_m18xr2
